TASK: Move test base classes into separate package that is required by flow as dev dependency
For backwards compatibility classes that extend the new namespace are left behind in the old place.
