+++ Neos.Flow.Testing/Classes/RequestHandler/SpyRequestHandler.php
@@ -1,0 +1,27 @@
+<?php
+namespace Neos\Flow\Testing\RequestHandler;
+
+use GuzzleHttp\Psr7\Response;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\RequestHandlerInterface;
+
+class SpyRequestHandler implements RequestHandlerInterface
+{
+
+    /**
+     * @var ServerRequestInterface
+     */
+    protected $handledRequest;
+
+    public function handle(ServerRequestInterface $request): ResponseInterface
+    {
+        $this->handledRequest = $request;
+        return new Response();
+    }
+
+    public function getHandledRequest(): ServerRequestInterface
+    {
+        return $this->handledRequest;
+    }
+}
