+++ Neos.Flow/Tests/FunctionalTestRequestHandler.php
@@ -11,11 +11,6 @@ namespace Neos\Flow\Tests;
  * source code.
  */
 
-use GuzzleHttp\Psr7\ServerRequest;
-use Neos\Flow\Annotations as Flow;
-use Neos\Flow\Core\Bootstrap;
-use Psr\Http\Message\ServerRequestInterface;
-
 /**
  * A request handler which boots up Flow into a basic runtime level and then returns
  * without actually further handling command line commands.
@@ -32,82 +27,8 @@ use Psr\Http\Message\ServerRequestInterface;
  *
  * @Flow\Proxy(false)
  * @Flow\Scope("singleton")
+ * @deprecated use \Neos\Flow\Testing\RequestHandler\FunctionalTestRequestHandler
  */
-class FunctionalTestRequestHandler implements \Neos\Flow\Http\HttpRequestHandlerInterface
+class FunctionalTestRequestHandler extends \Neos\Flow\Testing\RequestHandler\FunctionalTestRequestHandler
 {
-    /**
-     * @var \Neos\Flow\Core\Bootstrap
-     */
-    protected $bootstrap;
-
-    /**
-     * @var ServerRequestInterface
-     */
-    protected $httpRequest;
-
-    /**
-     * Constructor
-     *
-     * @param \Neos\Flow\Core\Bootstrap $bootstrap
-     */
-    public function __construct(Bootstrap $bootstrap)
-    {
-        $this->bootstrap = $bootstrap;
-    }
-
-    /**
-     * This request handler can handle requests in Testing Context.
-     *
-     * @return boolean If the context is Testing, true otherwise false
-     */
-    public function canHandleRequest()
-    {
-        return $this->bootstrap->getContext()->isTesting();
-    }
-
-    /**
-     * Returns the priority - how eager the handler is to actually handle the
-     * request.
-     *
-     * As this request handler can only be used as a preselected request handler,
-     * the priority for all other cases is 0.
-     *
-     * @return integer The priority of the request handler.
-     */
-    public function getPriority()
-    {
-        return 0;
-    }
-
-    /**
-     * Handles a command line request
-     *
-     * @return void
-     */
-    public function handleRequest()
-    {
-        $sequence = $this->bootstrap->buildRuntimeSequence();
-        $sequence->invoke($this->bootstrap);
-    }
-
-    /**
-     * @param ServerRequestInterface $request
-     */
-    public function setHttpRequest(ServerRequestInterface $request): void
-    {
-        $this->httpRequest = $request;
-    }
-
-    /**
-     * Returns the currently processed HTTP request
-     *
-     * @return ServerRequestInterface
-     */
-    public function getHttpRequest(): ServerRequestInterface
-    {
-        if ($this->httpRequest === null) {
-            $this->httpRequest = ServerRequest::fromGlobals();
-        }
-        return $this->httpRequest;
-    }
 }
