+++ Neos.Flow/Tests/Unit/Http/Fixtures/SpyRequestHandler.php
@@ -1,27 +1,9 @@
 <?php
 namespace Neos\Flow\Tests\Unit\Http\Fixtures;
 
-use GuzzleHttp\Psr7\Response;
-use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\ServerRequestInterface;
-use Psr\Http\Server\RequestHandlerInterface;
-
-class SpyRequestHandler implements RequestHandlerInterface
+/**
+ * @deprecated use \Neos\Flow\Testing\RequestHandler\SpyRequestHandler
+ */
+class SpyRequestHandler extends \Neos\Flow\Testing\RequestHandler\SpyRequestHandler
 {
-
-    /**
-     * @var ServerRequestInterface
-     */
-    protected $handledRequest;
-
-    public function handle(ServerRequestInterface $request): ResponseInterface
-    {
-        $this->handledRequest = $request;
-        return new Response();
-    }
-
-    public function getHandledRequest(): ServerRequestInterface
-    {
-        return $this->handledRequest;
-    }
 }
