+++ Neos.Flow/Tests/Unit/Validation/Validator/BooleanValueValidatorTest.php
@@ -13,8 +13,6 @@ namespace Neos\Flow\Tests\Unit\Validation\Validator;
 
 use Neos\Flow\Validation\Validator\BooleanValueValidator;
 
-require_once('AbstractValidatorTestcase.php');
-
 /**
  * Testcase for the true validator
  */
