+++ Neos.Flow/Tests/Unit/Validation/Validator/CountValidatorTest.php
@@ -13,8 +13,6 @@ namespace Neos\Flow\Tests\Unit\Validation\Validator;
 
 use Neos\Flow\Validation\Validator\CountValidator;
 
-require_once('AbstractValidatorTestcase.php');
-
 /**
  * Testcase for the count validator
  *
