+++ Neos.Flow/Tests/Unit/Validation/Validator/NumberRangeValidatorTest.php
@@ -13,8 +13,6 @@ namespace Neos\Flow\Tests\Unit\Validation\Validator;
 
 use Neos\Flow\Validation\Validator\NumberRangeValidator;
 
-require_once('AbstractValidatorTestcase.php');
-
 /**
  * Testcase for the number range validator
  */
