+++ Neos.Flow/Tests/Unit/Validation/Validator/StringLengthValidatorTest.php
@@ -14,8 +14,6 @@ namespace Neos\Flow\Tests\Unit\Validation\Validator;
 use Neos\Flow\Validation\Exception\InvalidValidationOptionsException;
 use Neos\Flow\Validation\Validator\StringLengthValidator;
 
-require_once('AbstractValidatorTestcase.php');
-
 /**
  * Testcase for the string length validator
  *
