+++ Neos.Flow/composer.json
@@ -54,9 +54,7 @@
         "egulias/email-validator": "^2.1.17 || ^3.0"
     },
     "require-dev": {
-        "vimeo/psalm": "~4.9.3",
-        "mikey179/vfsstream": "^1.6.1",
-        "phpunit/phpunit": "~9.1"
+        "neos/flow-testing": "*"
     },
     "replace": {
         "symfony/polyfill-php70": "*",
