+++ composer.json
@@ -34,7 +34,10 @@
         "egulias/email-validator": "^2.1.17 || ^3.0",
         "typo3fluid/fluid": "^2.7.0",
         "guzzlehttp/psr7": "^1.7, !=1.8.0",
-        "ext-mbstring": "*"
+        "ext-mbstring": "*",
+        "mikey179/vfsstream": "^1.6.1",
+        "phpunit/phpunit": "~9.1",
+        "vimeo/psalm": "~4.9.3"
     },
     "replace": {
         "symfony/polyfill-php70": "*",
@@ -47,6 +50,7 @@
         "neos/error-messages": "self.version",
         "neos/flow-log": "self.version",
         "neos/flow": "self.version",
+        "neos/flow-testing": "self.version",
         "neos/fluid-adaptor": "self.version",
         "neos/http-factories": "self.version",
         "neos/kickstarter": "self.version",
@@ -86,6 +90,9 @@
             "Neos\\Flow\\": [
                 "Neos.Flow/Classes"
             ],
+            "Neos\\Flow\\Testing\\": [
+                "Neos.Flow.Testing/Classes"
+            ],
             "Neos\\Flow\\Core\\Migrations\\": [
                 "Neos.Flow/Scripts/Migrations"
             ],
@@ -117,11 +124,6 @@
         "neos": {
             "warning": "AUTOGENERATED FILE, ONLY MODIFY THE .composer.json IN THIS DIRECTORY AND RUN THE COMPOSER MANIFEST MERGER SCRIPT PROVIDED BY THE BUILD ESSENTIALS PACKAGE."
         }
-    },
-    "require-dev": {
-        "mikey179/vfsstream": "^1.6.1",
-        "phpunit/phpunit": "~9.1",
-        "vimeo/psalm": "~4.9.3"
     },
     "autoload-dev": {
         "psr-4": {
